requirements.txt
azure.cosmos
datetime
